+++ cmp-worker-store5/src/commonMain/kotlin/io/github/mobilebytelabs/worker/store5/koin/WorkStore5KoinModule.kt
@@ -12,7 +12,11 @@ import org.koin.dsl.module
  * ```kotlin
  * startKoin {
  *     modules(
- *         workKoinModule(config = WorkerConfig(), workers = workerRegistry { /* ... */ }),
+ *         workKoinModule(
+ *             config = WorkerConfig(),
+ *             workers = workerRegistry { /* ... */ },
+ *             factory = desktopWorkManagerFactory(), // or platform factory of your choice
+ *         ),
  *         workStore5KoinModule,
  *         appModule,
  *     )